add reconnect function for sse streams on client

+++ client_src/js/services/client-worker.service.js
@@ -4,6 +4,7 @@ let worker = null;
 let savedWorkerEventFunction = null;
 let currentFileURL = null;
 const workerResource = "/assets/dist/sse.client.min.js";
+let ignoreFirstMessage = true;
 
 function webWorkerFunction(url, workerEventFunction) {
   worker = new Worker(workerResource);
@@ -17,10 +18,11 @@ function webWorkerFunction(url, workerEventFunction) {
 
 function handleVisibilityChange() {
   if (document.visibilityState === "hidden") {
-    if (worker !== null) {
+    if (worker !== null && !ignoreFirstMessage) {
       console.log("Screen Abandonded, closing web worker...");
       worker.terminate();
       worker = null;
+      ignoreFirstMessage = false;
     }
   } else {
     if (worker === null) {
